requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
sqlalchemy==2.0.36
mysql-connector-python==9.1.0
python-dotenv==1.0.1